requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-spectacular>=0.26.0
djangorestframework-simplejwt>=5.2.0 
redis>=4.5.0
celery>=5.3.0
python-dotenv>=1.0.0
requests>=2.31.0
django-cors-headers>=4.5.0